+++ lib/sensu/client/process.rb
@@ -297,15 +297,20 @@ module Sensu
       # using a calculated execution splay. The timers are stored in
       # the timers hash under `:run`, so they can be cancelled etc.
       # Check definitions are duplicated before processing them, in
-      # case they are mutated. The check `:issued` timestamp is set
-      # here, to mimic check requests issued by a Sensu server.
+      # case they are mutated. A check will not be executed if it is
+      # subdued. The check `:issued` timestamp is set here, to mimic
+      # check requests issued by a Sensu server.
       #
       # @param checks [Array] of definitions.
       def schedule_checks(checks)
         checks.each do |check|
           execute_check = Proc.new do
-            check[:issued] = Time.now.to_i
-            process_check_request(check.dup)
+            unless check_subdued?(check)
+              check[:issued] = Time.now.to_i
+              process_check_request(check.dup)
+            else
+              @logger.info("check execution was subdued", :check => check)
+            end
           end
           execution_splay = testing? ? 0 : calculate_execution_splay(check)
           interval = testing? ? 0.5 : check[:interval]
@@ -318,15 +323,16 @@ module Sensu
 
       # Setup standalone check executions, scheduling standard check
       # definition and check extension executions. Check definitions
-      # and extensions with `:standalone` set to `true` will be
-      # scheduled by the Sensu client for execution.
+      # and extensions with `:standalone` set to `true`, have a
+      # integer `:interval`, and do not have `:publish` set to `false`
+      # will be scheduled by the Sensu client for execution.
       def setup_standalone
         @logger.debug("scheduling standalone checks")
         standard_checks = @settings.checks.select do |check|
-          check[:standalone]
+          check[:standalone] && check[:interval].is_a?(Integer) && check[:publish] != false
         end
         extension_checks = @extensions.checks.select do |check|
-          check[:standalone] && check[:interval].is_a?(Integer)
+          check[:standalone] && check[:interval].is_a?(Integer) && check[:publish] != false
         end
         schedule_checks(standard_checks + extension_checks)
       end