+++ lib/sensu/utilities.rb
@@ -118,5 +118,69 @@ module Sensu
       end
       [substituted, unmatched_tokens]
     end
+
+    # Determine if the current time falls within a time window. The
+    # provided condition must have a `:begin` and `:end` time, eg.
+    # "11:30:00 PM", or `false` will be returned.
+    #
+    # @param condition [Hash]
+    # @option condition [String] :begin time.
+    # @option condition [String] :end time.
+    # @return [TrueClass, FalseClass]
+    def in_time_window?(condition)
+      if condition.has_key?(:begin) && condition.has_key?(:end)
+        begin_time = Time.parse(condition[:begin])
+        end_time = Time.parse(condition[:end])
+        if end_time < begin_time
+          if Time.now < end_time
+            begin_time = Time.parse("12:00:00 AM")
+          else
+            end_time = Time.parse("11:59:59 PM")
+          end
+        end
+        Time.now >= begin_time && Time.now <= end_time
+      else
+        false
+      end
+    end
+
+    # Determine if time window conditions for one or more days of the
+    # week are met. If a day of the week is provided, it can provide
+    # one or more conditions, each with a `:begin` and `:end` time,
+    # eg. "11:30:00 PM", or `false` will be returned.
+    #
+    # @param conditions [Hash]
+    # @option conditions [String] :days of the week.
+    # @return [TrueClass, FalseClass]
+    def in_time_windows?(conditions)
+      in_window = false
+      window_days = conditions[:days] || {}
+      if window_days[:all]
+        in_window = window_days[:all].any? do |condition|
+          in_time_window?(condition)
+        end
+      end
+      current_day = Time.now.strftime("%A").downcase.to_sym
+      if !in_window && window_days[current_day]
+        in_window = window_days[current_day].any? do |condition|
+          in_time_window?(condition)
+        end
+      end
+      in_window
+    end
+
+    # Determine if a check is subdued, by conditions set in the check
+    # definition. If any of the conditions are true, without an
+    # exception, the check is subdued.
+    #
+    # @param check [Hash] definition.
+    # @return [TrueClass, FalseClass]
+    def check_subdued?(check)
+      if check[:subdue]
+        in_time_windows?(check[:subdue])
+      else
+        false
+      end
+    end
   end
 end